+++ src/app/components/CopyButton/index.js
@@ -29,7 +29,7 @@ export default class CopyButton extends Component {
       this.setState({
         buttonText: this.defaultButtonText
       })
-    }, 2000)
+    }, 1000)
   }
 
   render () {